+++ js/big-picture.js
@@ -1,14 +1,14 @@
 import { isEscapeKey } from './util';
-import { arrayPic } from './data-generate.js';
+import { arrayPictures } from './data-generate.js';
+import { displayComments } from './display-comments.js';
 const bodyElement = document.querySelector('body');
 const bigPictureElement = document.querySelector('.big-picture');
 const bigPictureCancelElement = document.querySelector('.big-picture__cancel');
 const picturesElement = document.querySelector('.pictures');
-const socialCommentsElement = bigPictureElement.querySelector('.social__comments');
 const commentsCountElement = bigPictureElement.querySelector('.social__comment-count');
 const commentsLoaderElement = bigPictureElement.querySelector('.comments-loader');
 
-const onEscKeydown = (evt) => {
+const onEscapeKey = (evt) => {
   if (isEscapeKey(evt)) {
     evt.preventDefault();
     // eslint-disable-next-line
@@ -22,64 +22,40 @@ const openBigPicture = () => {
   bodyElement.classList.add('modal-open');
   commentsCountElement.classList.add('hidden');
   commentsLoaderElement.classList.add('hidden');
-  document.addEventListener('keydown', onEscKeydown);
+  document.addEventListener('keydown', onEscapeKey);
 };
 
 //* Функция закрытия бигфото
 const closeBigPicture = () => {
   bigPictureElement.classList.add('hidden');
   bodyElement.classList.remove('modal-open');
-  document.removeEventListener('keydown', onEscKeydown);
-};
-
-//* создание коментария
-const createComment = ({ avatar, name, message }) => {
-  const socialCommentElement = bigPictureElement.querySelector('.social__comment').cloneNode(true);
-  const socialPicture = socialCommentElement.querySelector('.social__picture');
-
-  socialPicture.src = avatar;
-  socialPicture.alt = name;
-  socialCommentElement.querySelector('.social__text').textContent = message;
-  return socialCommentElement;
-};
-
-//* создание списка коментариев текущей мииатюры
-const displayComments = (comments) => {
-  const fragment = document.createDocumentFragment();
-
-  comments.forEach((comment) => {
-    const commentEl = createComment(comment);
-    fragment.appendChild(commentEl);
-  });
-
-  socialCommentsElement.innerHTML = '';
-  socialCommentsElement.appendChild(fragment);
+  document.removeEventListener('keydown', onEscapeKey);
 };
 
 //*получаем данные бигфото
 const getBigPhoto = (photoId) => {
-  const picData = arrayPic[photoId];
+  const pictureData = arrayPictures[photoId];
 
   bigPictureElement.dataset.photoId = photoId; // Устанавливаем data-photo-id
-  bigPictureElement.querySelector('.big-picture__img img').src = picData.url;
-  bigPictureElement.querySelector('.likes-count').textContent = picData.likes;
-  bigPictureElement.querySelector('.social__caption').textContent = picData.description;
-  displayComments(picData.comments);
+  bigPictureElement.querySelector('.big-picture__img img').src = pictureData.url;
+  bigPictureElement.querySelector('.likes-count').textContent = pictureData.likes;
+  bigPictureElement.querySelector('.social__caption').textContent = pictureData.description;
+  displayComments(pictureData.comments);
 };
 
 
 //* функция при клике на миниатюру
-const clickPhoto = (evt) => {
+const onClickPhoto = (evt) => {
   const thumbnail = evt.target.closest('.picture');
 
   if (thumbnail) {
-    const currentPic = thumbnail.querySelector('.picture__img').dataset.photoId;
-    getBigPhoto(currentPic);
+    const currentPicture = thumbnail.querySelector('.picture__img').dataset.photoId;
+    getBigPhoto(currentPicture);
     openBigPicture();
   }
 };
 
 //* обработчик событий для миниатюр
-picturesElement.addEventListener('click', clickPhoto);
+picturesElement.addEventListener('click', onClickPhoto);
 //* обработчик события для крестика
 bigPictureCancelElement.addEventListener('click', closeBigPicture);
